Modify activity comment serializer to populate related models
Used the depth option to bring in related models but had problems
saving. This modifies the activity comment and its replies to
bring in their related model and modifies their viewsets to allow
saving of the activity comment and its replies.
Also modifies the comment view and adds replies to view. Still
a WIP.