Add the students habit to child view
Includes the students habit from the habits module to the child view.
Also modifies the default behaviour of the child view which was by
default when a user visits (or refreshes) the any view in the child
module, the user is taken to see the current activity. Now the user
is first taken to view the current activity, but a refresh on any
other view keeps the user on the same view.