Add 'more' views to the child module
Adds a view to the menu named 'more' in the child module
The more page comprises of links to - the childs' profile,
classes/groups, subjects they've taken and results. The profile,
classes/groups and subjects have all been added to the view and
can be accessed via the links, i.e these views can be viewed
within the child module.