Implement a full functional profile view
Adds 2 views for users to view and edit their profiles. Also implements
session logging out from these views. Provides sub views on the view
page for users to see their relative details like contact, about,
medicals and relatives.
What is yet to be implemented is a filtered view for the medicals as
this subview should only be displayed for students.