Initialize main site build
The main site contains pages where both authenticated and non
authenticated users can view. The main site basically displays
information about what kidkrumbs is about. It contains a home page
which is used as an advertising medium to advertise schools; an about
page detailing about kidkrumbs, a contact page and a schools page that
can be used to search for and view schools.
This commit implements a scafold for all pages (except the home page) and routes.